ai-accessibility-assistant-main/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

sqlalchemy>=2.0.35

textstat==0.7.3
spacy>=3.8.0

gtts==2.5.1

openai>=1.0.0

python-dotenv==1.0.1
keybert==0.9.0
google-genai>=1.0.0

typer==0.9.0
click==8.1.7
python-docx==1.1.2
nltk==3.9.1
pypdf>=4.0.0
python-multipart>=0.0.9